requirements.txt
# Core dependencies
gym-super-mario-bros==7.4.0
nes-py==8.2.1

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0

# Reinforcement Learning
stable-baselines3[extra]>=2.1.0
gymnasium>=0.29.0
shimmy[gym-v21]>=1.3.0  # For gym 0.21 compatibility with gymnasium

# Visualization and Monitoring
tensorboard>=2.14.0
matplotlib>=3.7.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
pandas>=2.0.0
wandb>=0.15.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0